requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
agno==1.1.12
groq==0.31.1
scikit-learn==1.3.2
google-api-python-client==2.100.0
pandas==2.1.3
Pillow==10.1.0
python-dotenv==1.0.0